requirements.txt
fastapi
uvicorn
speechrecognition
pydub
google-generativeai
edge-tts
python-dotenv
